Modernize OSX platform code in gmgen
While here clean up pcgen version of the same code.
